requirements.txt
cryptography==3.4.8
requests==2.26.0
snowflake_sqlalchemy==1.3.3
